requirements.txt
asgiref==3.6.0
beautifulsoup4==4.11.2
bs4==0.0.1
certifi==2022.12.7
charset-normalizer==3.0.1
Django==4.1.7
django-ckeditor==6.5.1
django-js-asset==2.0.0
djangorestframework==3.14.0
idna==3.4
Pillow==9.4.0
pytz==2023.3
requests==2.28.2
soupsieve==2.4
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.14
